requirements.txt
dash
plotly
pandas
yfinance
pandas_ta
vaderSentiment
python-dotenv
requests
beautifulsoup4
scikit-learn
joblib
gunicorn